pyproject.toml
[tool.poetry]
name = "finance-api"
version = "0.1.0"
description = ""
authors = ["Felipe Carvalho <felipe.carvalho07@gmail.com>"]

[tool.black]
line-length = 91
target-version = ["py38"]

[tool.isort]
line_length = 91
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
force_grid_wrap = 0
known_local_folder = "accounts,authy"
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
default_section = "THIRDPARTY"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "finance.settings"
addopts = "-v -x --cov=authy --cov-report=term-missing --ignore=stocks"
testpaths = [
    "accounts",
    "authy",
]


[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
djangorestframework = "^3.12.4"
prettyconf = "^2.2.1"
django-cors-headers = "^3.8.0"
drf-yasg = "^1.20.0"
factory-boy = "^3.2.0"
freezegun = "^1.1.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = "^21.8b0"
pytest = "^6.2.5"
pytest-cov = "^2.12.1"
pytest-django = "^4.4.0"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
